requirements/base.txt
Django==1.8.4
django-crispy-forms==1.5.2
django-registration==1.0
psycopg2==2.6.1
Unipath==1.1
wheel==0.24.0
